+++ lib/event_tracer/prometheus_logger.rb
@@ -1,16 +1,4 @@
 module EventTracer
-  module PrometheusHelper
-    module_function
-
-    def register_prometheus_metrics(prometheus_registry, name, labels)
-      prometheus_registry.counter(
-        name.to_sym,
-        docstring: "A counter for #{name}",
-        labels: labels
-      )
-    end
-  end
-
   class PrometheusLogger < BasicDecorator
 
     SUPPORTED_METRIC_TYPES = {
